Allocate memory for image cache once. (#1508)
* Allocate memory for image cache once. Image size should be size_t. Check for nullptr to see if cache filled.

* added to changelog

* Added channel and polarization indices to performance log for image cache fill

* Moved cache allocation back to fill function, but made it conditional. Fixed bad logic in cache validity checks.

* restored original constructor code

+++ src/Frame/Frame.h
@@ -296,7 +296,7 @@ protected:
     ContourSettings _contour_settings;
 
     // Image data cache and mutex
-    long long int _image_cache_size;
+    size_t _image_cache_size;
     std::unique_ptr<float[]> _image_cache;
     bool _image_cache_valid;       // cached image data is valid for current z and stokes
     queuing_rw_mutex _cache_mutex; // allow concurrent reads but lock for write
